Several tests are failing on master HEAD
* Fixes https://github.com/eclipse-jdt/eclipse.jdt.core/issues/2661
